refactor(aptu-core): split registry.rs into modular structure with provider constants
Split monolithic registry.rs (727 lines) into separate modules for improved maintainability:
- consts.rs: Provider name constants (PROVIDER_ANTHROPIC, PROVIDER_OPENROUTER, PROVIDER_GEMINI, PROVIDER_GROQ, PROVIDER_CEREBRAS, PROVIDER_ZENMUX, PROVIDER_ZAI)
- config.rs: Static ProviderConfig entries and lookup functions (get_provider, all_providers) with unit tests
- parsing.rs: CachedModelRegistry HTTP fetch/parse/cache logic with model parsing tests
- mod.rs: Facade module with pub use re-exports maintaining backward compatibility

All 14 existing tests relocated alongside implementation code and passing. No changes to provider api_url, api_key_env, or default model values. Public API preserved.

Fixes #1481
